integer energy to prevent roundoff errors change mating - child energy comes from mum only, bugs can control probability of mating, and every bug gets a chance (rather than picking two on any one square) change defaults so that a smaller population results (more updates per sec) warn on load fail
git-svn-id: https://qtbugs.googlecode.com/svn/trunk@7 4b5036b8-5fc7-11de-a55d-e76e375ae706

+++ energy_node.cpp
@@ -1,6 +1,6 @@
 #include "energy_node.h"
 #include <QBrush>
 
-EnergyNode::EnergyNode(float e, int x, int y): QPoint(x, y), energy(e) {
+EnergyNode::EnergyNode(int e, int x, int y): QPoint(x, y), energy(e) {
 }
